add-raw-buffer-pixmap-support: implementation
Adding an interface for taking in a raw buffer of pixels
for a specific format (i.e. unsigned char) which will then render
onto the paint surface.

In addition an example has been added in order to test out the API,
and give a reference as to how to use it.