Fix the window limits calculation for GTK w/ CSD
When CSD is enabled (which is the default for GTK on Wayland), the
titlebar height is now included as part of the window, whereas the
caller would have intended for the given limits to only cover the
drawable area.

Unfortunately, GTK doesn't seem to provide any particular APIs for
accessing the default titlebar, and trying to use a custom GtkHeaderBar
but intelligently drop it when needed would be quite a mess. Thus, the
simplest solution is just to measure the inner widget's size and use
the difference in heights to adjust the limits.