Fix the maximum radius for canvas::round_rect
Prior to this change, using a radius that was too high would result in
the arcs at the corners extending out too far and crossing one another.

For perfectly rounded ends to a rectangle, the circumference of the
circular arcs must be min(width, height). The radius should be half of
that circumference.